Basic admin and user roles in most of the functions

+++ accounts.py
@@ -17,6 +17,21 @@ class Accounts(db.Model):
         self.age = age
 
 
+def logged_in():
+    if user_id() > 0:
+        return True
+    flash("Toiminnon käyttäminen vaatii kirjautumisen.")
+    return False
+
+def is_admin():
+    sql = "SELECT access FROM accounts WHERE id=:user"
+    result = db.session.execute(sql, {"user":user_id()})
+    access = result.fetchone()[0]
+    if access == "admin":
+        return True
+    #flash("Toiminnon käyttäminen vaatii pääkäyttäjän oikeudet.")
+    return False
+
 def login(username, password):
     sql = "SELECT id, password, name, age FROM accounts WHERE username=:username"
     result = db.session.execute(sql, {"username":username})
@@ -40,9 +55,15 @@ def logout():
     del session["user_id"]
     del session["age"]
 
-def register(name, username, password, age):
-    hash_value = generate_password_hash(password)
+def register(name, username, password, password2, age):
+    if password != password2:
+        flash("Anna sama salasana kahdesti.")
+        return False
+    if not 0 < int(age) < 130:
+        flash("Anna kelvollinen ikä.")
+        return False
     try:
+        hash_value = generate_password_hash(password)
         sql = "INSERT INTO accounts (name, username, password, age) VALUES " \
             "(:name, :username, :password, :age)"
         db.session.execute(sql, {"name":name, "username":username, "password":hash_value, "age":age})
@@ -52,18 +73,26 @@ def register(name, username, password, age):
     return login(username, password)
 
 def get_account(id):
+    if not is_admin():
+        id = user_id()
     sql = "SELECT id, name, username, password, age FROM accounts WHERE id=:id"
     result = db.session.execute(sql, {"id":id})
     account = result.fetchone()
     return account
 
 def get_accounts():
+    if user_id() == 0:
+        return []
     sql = "SELECT id, name, username, password, age FROM accounts"
     result = db.session.execute(sql, {"id":id})
     acc_list = result.fetchall()
     return acc_list
 
 def update(id, new_name, new_username, new_age):
+    id = user_id()
+    if not 0 < int(new_age) < 130:
+        flash("Anna kelvollinen ikä.")
+        return False
     try:
         account = Accounts.query.get(id)
         account.name = new_name
@@ -72,13 +101,56 @@ def update(id, new_name, new_username, new_age):
         db.session.commit() 
         return True
     except:
+        flash("Jokin meni pieleen.")
+        return False
+
+def change_password(id, old_password, password, password2):
+    user = user_id()
+    if user != int(id):
+        flash("Ei oikeuksia muuttaa salasanaa.")
+        return False
+    if password != password2:
+        flash("Anna uusi salasana kahdesti.")
+        return False
+    else:
+        sql = "SELECT password FROM accounts WHERE id=:id"
+        result = db.session.execute(sql, {"id":id})
+        user = result.fetchone()
+        if not check_password_hash(user[0], old_password):
+            flash("Tarkista vanha salasana.")
+            return False
+        else:
+            try:
+                account = Accounts.query.get(id)
+                account.password = generate_password_hash(password)
+                db.session.commit() 
+                flash("Salasana on nyt vaihdettu.")
+                return True
+            except:
+                return False
+
+def reset_password(id, username):
+    if not is_admin():
+        flash("Ei oikeuksia salasanan vaihtoon.")
+        return False
+    try:
+        account = Accounts.query.get(id)
+        account.password = generate_password_hash(username)
+        db.session.commit()
+        flash("Salasana on nyt sama kuin käyttäjätunnus.")
+        return True
+    except:
         return False
 
 def delete_account(id):
+    if user_id() != int(id) and not is_admin():
+        flash("Ei oikeuksia käyttäjätunnuksen poistamiseen.")
+        return False
     try:
         account = Accounts.query.get(id)
         db.session.delete(account)
         db.session.commit()
+        logout()
         return True
     except:
         return False
